refactor move rule test
